(Nominally) move color_management to cms/
It still depends on color_encoding_internal.h
