+++ lib/jxl/color_management_test.cc
@@ -3,8 +3,6 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-#include "lib/jxl/color_management.h"
-
 #include <jxl/cms_interface.h>
 #include <stdint.h>
