+++ lib/jxl/enc_butteraugli_comparator.cc
@@ -8,7 +8,6 @@
 #include <algorithm>
 #include <vector>
 
-#include "lib/jxl/color_management.h"
 #include "lib/jxl/enc_image_bundle.h"
 
 namespace jxl {
