+++ lib/jxl/enc_comparator.cc
@@ -11,7 +11,6 @@
 #include <algorithm>
 
 #include "lib/jxl/base/compiler_specific.h"
-#include "lib/jxl/color_management.h"
 #include "lib/jxl/enc_gamma_correct.h"
 #include "lib/jxl/enc_image_bundle.h"
