+++ lib/jxl/image_bundle.cc
@@ -11,7 +11,6 @@
 #include "lib/jxl/base/byte_order.h"
 #include "lib/jxl/base/padded_bytes.h"
 #include "lib/jxl/base/printf_macros.h"
-#include "lib/jxl/color_management.h"
 #include "lib/jxl/fields.h"
 
 namespace jxl {
