+++ lib/jxl/image_metadata.cc
@@ -12,7 +12,6 @@
 #include "lib/jxl/base/byte_order.h"
 #include "lib/jxl/base/padded_bytes.h"
 #include "lib/jxl/cms/opsin_params.h"
-#include "lib/jxl/color_management.h"
 #include "lib/jxl/fields.h"
 #include "lib/jxl/frame_header.h"
 #include "lib/jxl/quantizer.h"
