+++ lib/jxl/jxl_cms.cc
@@ -37,9 +37,9 @@
 #include "lib/jxl/base/printf_macros.h"
 #include "lib/jxl/base/span.h"
 #include "lib/jxl/base/status.h"
+#include "lib/jxl/cms/color_management.h"  // MaybeCreateProfile CIEXYZFromWhiteCIExy ExtraTF
 #include "lib/jxl/cms/transfer_functions-inl.h"
 #include "lib/jxl/color_encoding_internal.h"
-#include "lib/jxl/color_management.h"
 #include "lib/jxl/field_encodings.h"
 #include "lib/jxl/image.h"
 #if JPEGXL_ENABLE_SKCMS
