+++ tools/ssimulacra2_main.cc
@@ -6,7 +6,6 @@
 #include <stdio.h>
 
 #include "lib/extras/codec.h"
-#include "lib/jxl/color_management.h"
 #include "lib/jxl/jxl_cms.h"
 #include "tools/file_io.h"
 #include "tools/ssimulacra2.h"
